IA_Generativa/Projeto 3/requirements.txt
# --- Google AI & RAG ---
google-genai
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
pypdf

# --- Frontend & Interface ---
streamlit
gradio

# --- Integrações & Web ---
requests
python-dotenv

# --- Utilitários ---
pandas
datetime